Change name of app file to base and move test file out of tests folder
